refactor db

+++ Back-end/src/main/java/domain/Abonament.java
@@ -3,14 +3,14 @@ package domain;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
-public class Abonament extends Entity<Long>{
+public class Abonament extends Entity<Integer>{
     private LocalDateTime dataIncepere;
     private LocalDateTime dataExpirare;
     private Double pret;
     private String tip;
     private Client codClient;
 
-    public Abonament(Long id,LocalDateTime dataIncepere, LocalDateTime dataExpirare, Double pret, String tip, Client codClient) {
+    public Abonament(Integer id,LocalDateTime dataIncepere, LocalDateTime dataExpirare, Double pret, String tip, Client codClient) {
         this.id = id;
         this.dataIncepere = dataIncepere;
         this.dataExpirare = dataExpirare;
@@ -19,12 +19,12 @@ public class Abonament extends Entity<Long>{
         this.codClient = codClient;
     }
     @Override
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
     @Override
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
