+++ Back-end/src/main/java/domain/User.java
@@ -3,14 +3,14 @@ package domain;
 import java.util.Objects;
 
 
-public class User extends Entity<Long> {
+public class User extends Entity<Integer> {
     private String nume;
     private String prenume;
     private String email;
     private String parola;
     private String CNP;
 
-    public User(Long id, String nume, String prenume, String email, String parola, String CNP) {
+    public User(Integer id, String nume, String prenume, String email, String parola, String CNP) {
         this.id = id;
         this.nume = nume;
         this.prenume = prenume;
@@ -20,12 +20,12 @@ public class User extends Entity<Long> {
     }
 
     @Override
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
     @Override
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
