+++ Back-end/src/main/java/repository/IRepoAbonament.java
@@ -2,5 +2,5 @@ package repository;
 
 import domain.Abonament;
 
-public interface IRepoAbonament extends IRepository<Long, Abonament>{
+public interface IRepoAbonament extends IRepository<Integer, Abonament>{
 }
