+++ Back-end/src/main/java/repository/IRepoClient.java
@@ -2,5 +2,5 @@ package repository;
 
 import domain.Client;
 
-public interface IRepoClient extends IRepository<Long, Client>{
+public interface IRepoClient extends IRepository<Integer, Client>{
 }
