+++ Back-end/src/main/java/repository/IRepoControlor.java
@@ -2,5 +2,5 @@ package repository;
 
 import domain.Controlor;
 
-public interface IRepoControlor extends IRepository<Long, Controlor>{
+public interface IRepoControlor extends IRepository<Integer, Controlor>{
 }
