+++ Back-end/src/main/java/repository/IRepoUser.java
@@ -2,6 +2,6 @@ package repository;
 
 import domain.User;
 
-public interface IRepoUser extends IRepository<Long, User>{
+public interface IRepoUser extends IRepository<Integer, User>{
     public User findOneByUsernameAndPassword(String username, String password);
 }
